requirements.txt
appdirs==1.4.4
black==20.8b1
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.0
peewee==3.13.3
psycopg2==2.8.6
python-dotenv==0.15.0
regex==2020.10.28
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
Werkzeug==1.0.1
